requirements.txt
accelerate
bitsandbytes
datasets
geonamescache
git+https://github.com/facebookresearch/segment-anything.git
googletrans==3.1.0a0
google-cloud-vision
img2dataset
linktransformer
matplotlib
onnx
onnxruntime
opencv-python
pandas
Pillow==9.5.0
plotly==5.18.0
pycocotools
pycountry==20.7.3
pytorch-lightning
seaborn
sentence-transformers
seqeval
simplejson
tensorboard
timm
torch==1.13.1
torchvision==0.14.1
transformers==4.31
# --upgrade diffusers[torch]
# -U kaleido